requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.34.0
pytz==2024.1
python-dateutil==2.9.0.post0